Update ai button style
