--- packages/web/src/components/ai-generated-button/AiGeneratedButton.module.css
@@ -1,3 +1,0 @@
-.aiGeneratedButtonContainer {
-  width: 100%;
-}
