+++ packages/web/src/pages/profile-page/components/desktop/ProfilePage.module.css
@@ -275,7 +275,7 @@
   margin-top: 32px;
 }
 
-.aiGeneratedButtonContainer {
+.aiGeneratedCalloutContainer {
   margin-top: 32px;
 }
